Ensure that scenario is cleaned up when the process is terminated with CTRL+C
Also, refactor the scenario code into its own class.
